+++ TowerDefenseGame/src/entities/livingbeings/IntermediateLizard.java
@@ -6,7 +6,7 @@ public class IntermediateLizard extends Lizard{
 
 	private static final long serialVersionUID = 1L;
 
-	public IntermediateLizard(int xPos, int yPos, int width, int height, int attack, int health, boolean friendly) {
+	public IntermediateLizard(int xPos, int yPos, int width, int height, int attack, double health, boolean friendly) {
 		super(xPos, yPos, width, height, attack, health, friendly);
 		if(friendly) {
 		    super.setWalkingPath(Path.IMAGE_INTERMEDIATE_LIZARD_PLAYER_WALK);
